+++ actionview/lib/action_view/helpers/tags/date_select.rb
@@ -40,7 +40,7 @@ module ActionView
         end
 
         def default_datetime(options)
-          return if options[:include_blank] || options[:prompt]
+          return if options[:disabled_blank] || options[:include_blank] || options[:prompt]
 
           case options[:default]
           when nil